requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-rename-app==0.1.6
djangorestframework==3.14.0
Markdown==3.4.2
Pillow==9.4.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
